requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
python-dateutil>=2.8.0
scikit-learn>=1.5.0
joblib>=1.3.0
matplotlib>=3.7.0
bottleneck>=1.3.6
flask>=3.0.0
flask-cors>=4.0.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4